requirements.txt
numpy
h5py==3.12.1
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
jsonschema==4.23.0

yt_dlp
pafy

Hypercorn==0.17.3
flask==3.0.3
flask_session==0.8.0
redis==5.0.8
paho-mqtt==2.1.0
pymongo==4.9.1
authlib==1.3.2
Flask-Cors==4.0.1
cryptography==43.0.1

shapely>=2.0.0
supervision>=0.21.0

psutil==6.0.0
gputil==1.4.0

#PyJWT==2.8.0
